Spark.CodeHelpers (spark v2.2.24)

Helpers for meta programming around code and code snippets

Functions

code_identifier(code)

@spec code_identifier(Macro.t()) :: binary()

Given a section of Elixir AST, generate a hash of the code to help with generating unique names.

lift_functions(value, key, caller)

@spec lift_functions(Macro.t(), atom(), Macro.Env.t()) :: Macro.t()

Lift anonymous and captured functions.

Acts as an AST transformer to allow these kinds of functions to be added in the AST:

In the case of captured functions, it ensures they are all captured remote functions (ie calls with both the module and function name present) - this often requires the definition of a new public function on the target module.

In the case of anonymous functions, it converts them into a new public function on the module and returns a (remote) function capture much like that of above.

prewalk(ast, fun)

Copy of Macro.prewalk/2 w/ a branch accumulator

prewalk(ast, acc, branch_acc, fun)

Copy of Macro.prewalk/3 w/ a branch accumulator

traverse(ast, acc, branch_acc, pre, post)

A copy of the corresponding Macro.traverse function that has a separate accumulator that only goes down each branch, only for pre