Spark.Dsl.Verifier behaviour (spark v2.2.24)

A verifier gets the dsl state and can return :ok or :error.

In a verifier, you can reference and depend on other modules without causing compile time dependencies.

Callbacks

verify(map)

@callback verify(map()) :: :ok | {:error, term()} | {:warn, String.t() | [String.t()]}

Functions

fetch_option(dsl_state, path, option)

See Spark.Dsl.Transformer.fetch_option/3.

get_entities(dsl_state, path)

See Spark.Dsl.Transformer.get_entities/2.

get_option(dsl_state, path, option)

See Spark.Dsl.Transformer.get_option/3.

get_option(dsl_state, path, option, default)

See Spark.Dsl.Transformer.get_option/4.

get_persisted(dsl, key)

See Spark.Dsl.Transformer.get_persisted/2.

get_persisted(dsl, key, default)

See Spark.Dsl.Transformer.get_persisted/3.