Spark.Igniter (spark v2.2.24)

Helpers for patching Spark DSLs.

Functions

add_extension(igniter, module, type, key, extension, singleton? \\ false)

@spec add_extension(Igniter.t(), module(), module(), atom(), module(), boolean()) ::
  Igniter.t()

Adds an extension to a DSL module.

get_option(igniter, module, path)

@spec get_option(Igniter.t(), module(), [atom()]) ::
  {Igniter.t(), {:ok, Macro.t()} | :error}

Gets an option at a given path within a DSL. We will attempt to expand literals using the environment at the path but this is only guaranteed to return the AST at that option, not necessarily a value.

Additionally, this only finds options set explicitly in the body of the resource, not by an extension.

has_extension(igniter, module, type, key, extension)

@spec has_extension(Igniter.t(), module(), module(), atom(), module()) ::
  {Igniter.t(), boolean()}

Returns {igniter, true} if the module has the extension, or {igniter, false} otherwise.

prepend_to_section_order(igniter, type, sections)

Prepends a new section or list of sections to the section order in a formatter configuration.

remove_extension(igniter, module, type, key, extension, singleton? \\ false)

@spec remove_extension(Igniter.t(), module(), module(), atom(), module(), boolean()) ::
  Igniter.t()

Removes an extension from a DSL module.

set_option(igniter, module, path, value, updater \\ &{:ok, &1})

@spec set_option(
  Igniter.t(),
  module(),
  dsl_path :: [atom()],
  value :: term(),
  (Sourceror.Zipper.t() ->
     {:ok, Sourceror.Zipper.t()} | {:error, term() | [term()]} | :error)
) :: Igniter.t()

Sets an option at a given path within in a DSL.

update_dsl(igniter, module, path, value, func)